Change tests to reflect planned refactor
